markdown_erb needs to sanitize html too
