+++ apps/dashboard/app/models/motd_formatter/base_formatter.rb
@@ -1,0 +1,14 @@
+module MotdFormatter
+  
+  class BaseFormatter
+    include ActionView::Helpers::SanitizeHelper
+
+    def safe_content(content)
+      if Configuration.motd_render_html?
+        content.html_safe
+      else
+        sanitize(content)
+      end
+    end
+  end
+end
