+++ apps/dashboard/app/models/motd_formatter/markdown_erb.rb
@@ -1,16 +1,19 @@
+# frozen_string_literal: true
+
 module MotdFormatter
   # Utility class for rendering Markdown MOTD files after ERB rendering them.
- class MarkdownErb
+  class MarkdownErb < BaseFormatter
     attr_reader :content, :title
 
     def initialize(motd_file)
       motd_file ||= MotdFile.new unless motd_file
       @title = motd_file.title
-      @content = OodAppkit.markdown.render(ERB.new(motd_file.content).result)
+      content = OodAppkit.markdown.render(ERB.new(motd_file.content).result)
+      @content = safe_content(content)
     end
-    
+
     def to_partial_path
-      "dashboard/motd_markdown"
+      'dashboard/motd_markdown'
     end
- end  
+  end
 end
